auth/requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-allauth[socialaccount]
django-unfold==0.20.1
psycopg2-binary==2.9.9
drf-spectacular==0.27.0
Pillow==10.0.1
celery
redis
requests
requests_oauthlib

#AI and NLP
openai>=1.14.0
langchain==0.1.4
python-dotenv==1.0.1